+++ packages/dashboards/live-shopfloor/src/views/Index.vue
@@ -35,6 +35,7 @@ export default {
   },
   async created() {
     this.setLoading(true);
+    await this.getShifts();
     await this.getBusinessTime();
     if (this.isLoaded) {
       await this.getMachines();
@@ -75,8 +76,21 @@ export default {
     },
   },
   methods: {
-    ...mapMutations('shopfloor', ['setLoading', 'setMachine']),
-    ...mapActions('shopfloor', ['getBusinessTime', 'getMachines']),
+    ...mapMutations('shopfloor', [
+      'setLoading',
+      'setMachine',
+      'setCurrentShift',
+      'setCurrentHour',
+      'setDisplayHour',
+      'setCurrentDate',
+    ]),
+    ...mapActions('shopfloor', [
+      'getBusinessTime',
+      'getMachines',
+      'getShifts',
+      'getShiftAvailableTime',
+      'fetchRejections',
+    ]),
     sseInit() {
       this.sseClient = new EventSource('/sse/asm');
       this.readyState = this.sseClient.readyState;
@@ -108,42 +122,107 @@ export default {
         this.sseInit();
       }
     },
-    setEventData(data) {
-      const {
-        elementName, machinename, planid, partname, qty, target, updatedAt, actm_sum: runtime, sctm,
-      } = data;
+    async setEventData(data) {
+      const { elementName } = data;
+      this.setTime(data);
+      const shiftWorkingTime = await this.getShiftAvailableTime();
+      const payload = {
+        ...data,
+        shiftWorkingTime,
+      };
       if (elementName === 'cycletime') {
-        this.machines.forEach((m, index) => {
-          if (m.machinename === machinename) {
-            const payload = {
-              ...m,
-              updatedat: updatedAt,
-            };
-            if (planid === m.planid) {
-              let productions = [...m.production];
-              const partIndex = m.production.findIndex((prod) => prod.partname === partname);
-              if (partIndex > -1) {
-                productions[partIndex].produced = qty;
-              } else {
-                productions = [...productions, {
-                  partname,
-                  produced: qty,
-                  rejected: 0,
-                  target,
-                }];
-              }
-              payload.production = productions;
-            } else {
-              payload.planid = planid;
-              payload.runtime = runtime;
-              payload.stdcycletime = sctm;
-            }
-            this.setMachine({ index, payload });
-          }
-        });
+        this.setProduction(payload);
+      } else if (elementName === 'rejection') {
+        await this.setRejection(payload);
+      } else if (elementName === 'downtime') {
+        this.setDowntime(payload);
       }
       // if element name rejection - fetch from rejection element
       // if element name downtime - update machine record
+    },
+    setProduction(data) {
+      const {
+        machinename, planid, partname, qty, cavity, updatedAt, actm_sum: runtime, sctm,
+      } = data;
+      this.machines.forEach((m, index) => {
+        if (m.machinename === machinename) {
+          const payload = {
+            ...m,
+            updatedAt: new Date(updatedAt).getTime(),
+          };
+          if (planid === m.planid) {
+            let productions = [...m.production];
+            const partIndex = m.production.findIndex((prod) => prod.partname === partname);
+            const target = Math.ceil((runtime / sctm) * cavity);
+            if (partIndex > -1) {
+              productions[partIndex].produced = qty;
+              productions[partIndex].target = target;
+            } else {
+              productions = [...productions, {
+                partname,
+                produced: qty,
+                rejected: 0,
+                target,
+              }];
+            }
+            payload.production = productions;
+          } else {
+            payload.planid = planid;
+            payload.runtime = runtime;
+            payload.stdcycletime = sctm;
+          }
+          this.setMachine({ index, payload });
+        }
+      });
+    },
+    async setRejection(data) {
+      const {
+        machinename, planid, partname,
+      } = data;
+      this.machines.forEach(async (m, index) => {
+        if (m.machinename === machinename) {
+          const payload = {
+            ...m,
+            updatedAt: new Date().getTime(),
+          };
+          if (planid === m.planid) {
+            const productions = [...m.production];
+            const partIndex = m.production.findIndex((prod) => prod.partname === partname);
+            if (partIndex > -1) {
+              productions[partIndex].rejected = await this.fetchRejections({
+                planId: planid,
+                part: partname,
+              });
+            }
+            payload.production = productions;
+          }
+          this.setMachine({ index, payload });
+        }
+      });
+    },
+    setDowntime(data) {
+      const {
+        machinename, actualdowntimestart,
+      } = data;
+      for (let i = 0; i < this.machines.length; i += 1) {
+        if (this.machines[i].machinename === machinename) {
+          const payload = {
+            ...this.machines[i],
+            updatedAt: new Date().getTime(),
+            downsince: actualdowntimestart,
+            downreason: '',
+          };
+          this.setMachine({ index: i, payload });
+        }
+      }
+    },
+    setTime(data) {
+      const {
+        shift, hour, displayHour,
+      } = data;
+      this.setCurrentShift(shift);
+      this.setCurrentHour(hour);
+      this.setDisplayHour(displayHour);
     },
   },
 };