+++ packages/infinity-modules/production-log/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.2.0-alpha.2](https://bitbucket.org/entrib/shopworx/compare/v4.2.0-alpha.1...v4.2.0-alpha.2) (2020-12-14)
+
+
+### Bug Fixes
+
+* **production-og:** fixed accepted qty calculation ([bd1ad0b](https://bitbucket.org/entrib/shopworx/commits/bd1ad0bc3c09895cbc6be027495d79a7f3f29288))
+
+
+
+
+
 # [4.2.0-alpha.1](https://bitbucket.org/entrib/shopworx/compare/v4.1.3...v4.2.0-alpha.1) (2020-12-08)